+++ contracts/interfaces/cover/ICoverPoolManager.sol
@@ -8,14 +8,14 @@ interface ICoverPoolManager is CoverPoolStructs {
     function owner() external view returns (address);
     function feeTo() external view returns (address);
     function poolTypes(
-        uint8 poolTypeId
+        uint16 poolTypeId
     ) external view returns (
         address poolImpl,
         address tokenImpl,
         address twapImpl
     );
     function volatilityTiers(
-        uint8 poolTypeId,
+        uint16 poolTypeId,
         uint16 feeTier,
         int16  tickSpread,
         uint16 twapLength
